==> Delay Coordinate Embedding, realtime prediction of Changzhou GPS trajectory/1. LoadDataAndWriteCSV.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('Route_11_bus_51938440.txt',encoding='utf-8')

df.shape

df['ACTDATETIME'] = df['ACTDATETIME'].astype('datetime64[ns]')

df.head()

#得把运营类型在2，3删除，就是在首末站的那些数据
db = df[df['BUSSID'].isin([4,5])]

### Data Check
选择工作日的数据，并输出查看，寻找noise filter的方式。
通过查看有很多在首末站的重复记录，按照Nonlinear Dynamics课程的思路，创建一个low frequency filter，只让低频信号通过。
高频就是terminal station的GPS重复出现高达50(自定义)次以上，可以认为是high frequency即noise。

db_1 = db[(db['ACTDATETIME'] >'2017-05-22 00:00:00') & (db['ACTDATETIME'] < '2017-05-27 00:00:00')]

db_2 = db_1.loc[:,['ACTDATETIME','LONGITUDE','LATITUDE','BUSSID','STATIONTYPE','STATIONSEQNUM']]

db_2.to_csv('check.csv',index = False)

## Noise filter 

#LonLat = db_2.loc[:,['LONGITUDE','LATITUDE']].round(6)
LonLat = db.loc[:,['LONGITUDE','LATITUDE']].round(6)

### Noise filter by columns

a = LonLat.iloc[:,0]
b = LonLat.iloc[:,1]
a_1 = a.groupby((a.shift() != a).cumsum())\
.filter(lambda x: len(x) <30)
b_1 = b.groupby((b.shift() != a).cumsum())\
.filter(lambda x: len(x) <30)

a_1.dropna(inplace=True)

b_1.dropna(inplace=True)

a_1.to_csv('before_DCE_for_route11_Bus51938440.txt',header=None, index=None, sep=' ', mode='w')

### Noise filter by Table with two operating direction
It's better to filter noise on the whole table, because filtering noise by comlumns may end up having columns with different length.

A side note: One still have to do filtering based on only one column, here I decided to do that by the longitude column.

bbb = db.loc[:,['ACTDATETIME','LONGITUDE','LATITUDE','BUSSID','ISARRLFT','STATIONTYPE','STATIONSEQNUM']].copy()

bbb

bbb = bbb.groupby((bbb.iloc[:,1].shift() != bbb.iloc[:,1]).cumsum())\
.filter(lambda x: len(x) <20)

bbb.loc[:,['LONGITUDE','LATITUDE']].to_csv('before_DCE_for_route11_Bus51938440.txt',header=None, index=None, sep=' ', mode='w')

bbb.to_csv('noise_filtered_Route11Bus51938440.csv',index=False)

### Noise filter by Table with only one direction
Since the result of Delay Coordinate Embedding(DEC) is not good as I expected to be, now I try to only use data from one direction to test its performance on DCE.

db_downstream = df[df['BUSSID']==4].loc[:,['ACTDATETIME','LONGITUDE','LATITUDE','BUSSID','STATIONTYPE','STATIONSEQNUM']] #select downstream data

db_downstream_filter = db_downstream.groupby((db_downstream.iloc[:,1].shift() != db_downstream.iloc[:,1]).cumsum())\
.filter(lambda x: len(x) <20)

db_downstream_filter.loc[:,['LONGITUDE','LATITUDE']].to_csv('before_DCE_for_route11_Bus51938440_downstream.txt',header=None, index=None, sep=' ', mode='w')
