==> Delay Coordinate Embedding, realtime prediction of Changzhou GPS trajectory/2. DelayCoordinateEmbedding.ipynb <==
#using Pkg
#Pkg.add("PyCall")
#PyCall下载下来后但是安装失败用下面的code
#Pkg.build("PyCall")

using DelimitedFiles
using DynamicalSystems
using PyPlot

pyversion

LongLat=float(open(readdlm,"before_DCE_for_route11_Bus51938440.txt"))

Long = LongLat[:,1]
Lat = LongLat[:,2]

Long

### parameter estimation

t_long = estimate_delay(Long, "mi_min", 0:1:400)

d_long = estimate_dimension(Long, t_long, 1:50, "fnn")

re_Long = reconstruct(Long,9,t_long)

t_lat = estimate_delay(Lat, "mi_min", 0:1:400)

d_lat = estimate_dimension(Lat, t_lat, 1:20, "fnn")

re_Lat = reconstruct(Lat,10,t_lat)

### write reconstructed data to csv file

writedlm( "reconstructed_Longitude.csv",  re_Long, ',')

writedlm( "reconstructed_Lat.csv",  re_Lat, ',')

### One direction DCE
not good

LongLat_downstream=float(open(readdlm,"before_DCE_for_route11_Bus51938440_downstream.txt"));

Long_downstream = LongLat_downstream[:,1];
Lat_downstream = LongLat_downstream[:,2];

t_long_ds = estimate_delay(Long_downstream, "mi_min", 0:1:400)

d_long_ds = estimate_dimension(Long_downstream, t_long_ds, 1:20, "fnn")
