==> Delay Coordinate Embedding, realtime prediction of Changzhou GPS trajectory/3. Unreconstructed_prediction.ipynb <==
import pandas as pd
import numpy as np

from sklearn.model_selection import KFold #cross_validation

from sklearn.linear_model import BayesianRidge, LinearRegression, ElasticNet  # 批量导入要实现的回归算法
from sklearn.svm import LinearSVR  # SVM中的回归算法
from sklearn.svm import SVR  # SVM中的回归算法
from sklearn.ensemble.gradient_boosting import GradientBoostingRegressor  # 集成算法
from sklearn.neighbors import KNeighborsRegressor #  knn
from sklearn.model_selection import cross_val_score  # 交叉检验
from sklearn.metrics import explained_variance_score, mean_absolute_error, mean_squared_error, r2_score  # 批量导入指标算法
from sklearn.model_selection import GridSearchCV 
import matplotlib.pyplot as plt  # 导入图形展示库
from sklearn.preprocessing import StandardScaler #引入标准化模块
import seaborn as sns
from sklearn.neural_network import MLPRegressor
from sklearn.utils.validation import check_array as check_arrays
from sklearn.model_selection import train_test_split

### Combine reconstructed LongLat data with full table created by step 1.

#read reconstructed longitude and latitude
re_Long = pd.read_csv('reconstructed_Longitude.csv',encoding='utf-8',header=None)
re_Lat = pd.read_csv('reconstructed_Lat.csv',encoding='utf-8',header=None)

# read original yet noise-filtered table
original_table = pd.read_csv('noise_filtered_Route11Bus51938440.csv',encoding='utf-8',header=None)

re_Long and re_lat don't have  the same length, so one needs to find the minimum of them before concat them to original table

min_len = min([len(re_Long),len(re_Lat)])

### prepare for the prediction dataset
From above, we know the min_len is the length of re_Lat

# original table should start here, at (len(original_table) - min_len)
table_1 = original_table.iloc[(len(original_table) - min_len):,:]

# slice re_Long in order to get the same length as re_Lat
re_Long = re_Long.tail(min_len) # get the last min_len rows of re_Long


## test prection based on unrecontructed data

# Do not run this one
def prediction_process(X,y):
    # 训练回归模型
    X_train_no_cv, X_test_no_cv, y_train_no_cv, y_test_no_cv = train_test_split(
    X, y, test_size=0.3, shuffle=False)
    
    n_folds = 6  # 设置交叉检验的次数
    model_knn = KNeighborsRegressor()  # 建立贝叶斯岭回归模型对象
    model_lr = LinearRegression()  # 建立普通线性回归模型对象
    model_ann = MLPRegressor(solver='lbfgs', alpha=1e-5,hidden_layer_sizes=(5, 2), random_state=1)  # 建立ANN
    model_svr = LinearSVR(max_iter=3000)  # 建立支持向量机回归模型对象
    model_gbr = GradientBoostingRegressor(learning_rate=0.06, n_estimators=600,max_depth=5, min_samples_leaf =60, 
                   min_samples_split =1000, subsample=0.8)  # 建立梯度增强回归模型对象
    model_names = ['KNN', 'KF', 'ANN', 'LS-SVM', 'GBRT']  # 不同模型的名称列表
    model_dic = [model_knn, model_lr, model_ann, model_svr, model_gbr]  # 不同回归模型对象的集合

    def mean_absolute_percentage_error(y_true, y_pred):
        y_true, y_pred = np.array(y_true), np.array(y_pred)
        return np.mean(np.abs((y_true - y_pred) / y_true)) * 100

    cv_score_list = []  # 交叉检验结果列表
    pre_y_list = []  # 各个回归模型预测的y值列表
    for model in model_dic:  # 读出每个回归模型对象
        
        kf = KFold(n_splits=n_folds)
        scores = []
        for train_index, test_index in kf.split(X):
            X_train, X_test = X[train_index], X[test_index]
            y_train, y_test = y[train_index], y[test_index]
            
            y_pred = model.fit(X_train,y_train).predict(X_test)
            
            y_diff = abs(y_test-y_pred)
            y_no_noise_loc = np.where(y_diff < 0.1)[0]
            y_test_filtered = y_test[y_no_noise_loc]
            y_pred_filtered = y_pred[y_no_noise_loc]
            scores.append(mean_absolute_error(y_test_filtered,y_pred_filtered))
        scores = np.array(scores)
        scores = scores.reshape([n_folds,])
        
        y_pred = model.fit(X_train_no_cv, y_train_no_cv).predict(X_test_no_cv)
        y_diff = abs(y_test_no_cv-y_pred)
        y_no_noise_loc = np.where(y_diff < 0.1)[0]
        y_test_filtered = y_test_no_cv[y_no_noise_loc]
        y_pred_filtered = y_pred[y_no_noise_loc]

        cv_score_list.append(scores)  # 将交叉检验结果存入结果列表
        pre_y_list.append(y_pred_filtered)  # 将回归训练中得到的预测y存入列表，要分训练集和预测集
    # 模型效果指标评估
    n_samples, n_features = X.shape  # 总样本量,总特征数
    model_metrics_name = [explained_variance_score, mean_absolute_error, mean_squared_error, r2_score]  # 回归评估指标对象集
    model_metrics_list = []  # 回归评估指标列表
    for i in range(5):  # 循环每个模型索引
        tmp_list = []  # 每个内循环的临时结果列表
        for m in model_metrics_name:  # 循环每个指标对象
            tmp_score = m(y_test_filtered, pre_y_list[i])  # 计算每个回归指标结果
            tmp_list.append(tmp_score)  # 将结果存入每个内循环的临时结果列表
        model_metrics_list.append(tmp_list)  # 将结果存入回归评估指标列表
    df1 = pd.DataFrame(cv_score_list, index=model_names)  # 建立交叉检验的数据框
    df2 = pd.DataFrame(model_metrics_list, index=model_names, columns=['ev', 'mae', 'mse', 'r2'])  # 建立回归指标的数据框
    print ('samples: %d \t features: %d' % (n_samples, n_features))  # 打印输出样本量和特征数量
    print (70 * '-')  # 打印分隔线
    print ('cross validation result:')  # 打印输出标题
    print (df1)  # 打印输出交叉检验的数据框
    print (70 * '-')  # 打印分隔线
    print ('regression metrics:')  # 打印输出标题
    print (df2)  # 打印输出回归指标的数据框
    print (70 * '-')  # 打印分隔线
    print ('short name \t full name')  # 打印输出缩写和全名标题
    print ('ev \t explained_variance')
    print ('mae \t mean_absolute_error')
    print ('mse \t mean_squared_error')
    print ('r2 \t r2')
    print (70 * '-')  # 打印分隔线
    
    return pre_y_list, y_test_filtered

def prediction_process(X,y):
    # 训练回归模型
    X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, shuffle=False) # don't shuffle the data
    
    model_knn = KNeighborsRegressor()  # 建立贝叶斯岭回归模型对象
    model_lr = LinearRegression()  # 建立普通线性回归模型对象
    model_ann = MLPRegressor(solver='lbfgs', alpha=1e-5,hidden_layer_sizes=(5, 2), random_state=1)  # 建立ANN
    model_svr = SVR(max_iter=3000)  # 建立支持向量机回归模型对象
    model_gbr = GradientBoostingRegressor(learning_rate=0.06, n_estimators=600,max_depth=5, min_samples_leaf =60, 
                   min_samples_split =1000, subsample=0.8)  # 建立梯度增强回归模型对象
    model_names = ['KNN', 'KF', 'ANN', 'LS-SVM', 'GBRT']  # 不同模型的名称列表
    model_dic = [model_knn, model_lr, model_ann, model_svr, model_gbr]  # 不同回归模型对象的集合

    def mean_absolute_percentage_error(y_true, y_pred):
        y_true, y_pred = np.array(y_true), np.array(y_pred)
        return np.mean(np.abs((y_true - y_pred) / y_true)) * 100

    pre_y_list = []  # 各个回归模型预测的y值列表
    for model in model_dic:  # 读出每个回归模型对象
        pre_y_list.append(model.fit(X_train, y_train).predict(X_test))  # 将回归训练中得到的预测y存入列表，要分训练集和预测集
    # 模型效果指标评估
    n_samples, n_features = X.shape  # 总样本量,总特征数
    model_metrics_name = [explained_variance_score, mean_absolute_error, mean_squared_error, r2_score]  # 回归评估指标对象集
    model_metrics_list = []  # 回归评估指标列表
    for i in range(5):  # 循环每个模型索引
        tmp_list = []  # 每个内循环的临时结果列表
        for m in model_metrics_name:  # 循环每个指标对象
            tmp_score = m(y_test, pre_y_list[i])  # 计算每个回归指标结果
            tmp_list.append(tmp_score)  # 将结果存入每个内循环的临时结果列表
        model_metrics_list.append(tmp_list)  # 将结果存入回归评估指标列表
    df2 = pd.DataFrame(model_metrics_list, index=model_names, columns=['ev', 'mae', 'mse', 'r2'])  # 建立回归指标的数据框
    print ('samples: %d \t features: %d' % (n_samples, n_features))  # 打印输出样本量和特征数量
    print (70 * '-')  # 打印分隔线
    print ('regression metrics:')  # 打印输出标题
    print (df2)  # 打印输出回归指标的数据框
    print (70 * '-')  # 打印分隔线
    print ('short name \t full name')  # 打印输出缩写和全名标题
    print ('ev \t explained_variance')
    print ('mae \t mean_absolute_error')
    print ('mse \t mean_squared_error')
    print ('r2 \t r2')
    print (70 * '-')  # 打印分隔线
    
    return pre_y_list, y_test

### one-step prediction

#### one-step for long

prediction_step = 1

X_Long = np.array(re_Long.iloc[:,-1].head(min_len-prediction_step)) # cause y has to shift prediction_step units用这一个reconstructed space中的state variable 预测下一个状态的original space中的state variable

need to reshap X, an array of dimension (number,)doesn't work for sklearn predictor

X_Long = X_Long.reshape([min_len-prediction_step,1]) 

y_Long = np.array(re_Long.iloc[:,-1].tail(min_len-prediction_step))
y_Long = y_Long.reshape([min_len-prediction_step,1]) 

pre_y_list, y_test = prediction_process(X_Long,y_Long)

np.set_printoptions(edgeitems=100)
y_test

pre_y_list[0]

y_test-pre_y_list[0]

test = abs(y_test-pre_y_list[0])

len(test[test>0.01])

len(test[test>0.1])

one-step error even for unreconstructed data is small.

output these data for visualization purpose.

y_no_noise_loc = np.where(test < 0.1)[0]
y_test_filtered = y_test[y_no_noise_loc].reshape(-1,1)
y_pred_filtered = pre_y_list[0][y_no_noise_loc]

np.c_[y_test_filtered, y_pred_filtered]

long_1_y = y_test
long_1_y_pred = pre_y_list[0]

mean_absolute_error(y_test_filtered,y_pred_filtered)

#### one-step for Lat

prediction_step = 1
X_Lat = np.array(re_Lat.iloc[:,-1].head(min_len-prediction_step))
X_Lat = X_Lat.reshape([min_len-prediction_step,1]) 
y_Lat = np.array(re_Lat.iloc[:,-1].tail(min_len-prediction_step))
y_Lat = y_Lat.reshape([min_len-prediction_step,1]) 

pre_y_list, y_test = prediction_process(X_Lat,y_Lat)

np.set_printoptions(edgeitems=100)
y_test

pre_y_list[0]

y_test-pre_y_list[0]

test = abs(y_test-pre_y_list[0])

len(test[test>0.01])

len(test[test>0.1])

y_no_noise_loc = np.where(test < 0.1)[0]
y_test_filtered = y_test[y_no_noise_loc].reshape(-1,1)
y_pred_filtered = pre_y_list[0][y_no_noise_loc]

np.c_[y_test_filtered, y_pred_filtered]

lat_1_y = y_test
lat_1_y_pred = pre_y_list[0]

mean_absolute_error(y_test_filtered,y_pred_filtered)

### Multi-step prediction

prediction_step = 60

X_Long = np.array(re_Long.iloc[:,-1].head(min_len-prediction_step)) # cause y has to shift prediction_step units用这一个reconstructed space中的state variable 预测下一个状态的original space中的state variable
X_Long = X_Long.reshape([min_len-prediction_step,1]) 
y_Long = np.array(re_Long.iloc[:,-1].tail(min_len-prediction_step))
y_Long = y_Long.reshape([min_len-prediction_step,1]) 

multi_pre_y_list, multi_y_test = prediction_process(X_Long,y_Long)

multi_y_test

multi_pre_y_list[0]

len(multi_pre_y_list[0])

multi_y_test-multi_pre_y_list[0]

test = abs(multi_y_test-multi_pre_y_list[0])

len(test[test>0.1])

len(test[test>0.01])

y_no_noise_loc = np.where(test < 0.1)[0]
y_test_filtered = multi_y_test[y_no_noise_loc].reshape(-1,1)
y_pred_filtered = multi_pre_y_list[0][y_no_noise_loc]

long_multi_y = multi_y_test
long_multi_y_pred = multi_pre_y_list[0]

mean_absolute_error(y_test_filtered,y_pred_filtered)

#### prediction test for latitude

prediction_step = 60
X_Lat = np.array(re_Lat.iloc[:,-1].head(min_len-prediction_step))
X_Lat = X_Lat.reshape([min_len-prediction_step,1]) 
y_Lat = np.array(re_Lat.iloc[:,-1].tail(min_len-prediction_step))
y_Lat = y_Lat.reshape([min_len-prediction_step,1]) 

multi_pre_y_list_Lat, multi_y_test_Lat = prediction_process(X_Lat,y_Lat)

multi_pre_y_list_Lat[0]

multi_y_test_Lat

abs(multi_pre_y_list_Lat[0]-multi_y_test_Lat)

test_Lat = abs(multi_pre_y_list_Lat[0]-multi_y_test_Lat)

len(test_Lat[test_Lat>0.1])

len(test_Lat[test_Lat>0.01])

y_no_noise_loc = np.where(test_Lat < 0.1)[0]
y_test_filtered = multi_y_test_Lat[y_no_noise_loc].reshape(-1,1)
y_pred_filtered = multi_pre_y_list_Lat[0][y_no_noise_loc]

lat_multi_y = multi_y_test_Lat
lat_multi_y_pred = multi_pre_y_list_Lat[0]

mean_absolute_error(y_test_filtered,y_pred_filtered)

误差主要存在于mae再0.1-0.01这个范围，但是embedding以后的效果比这个好很多

### Output for Visualization

pd.DataFrame(np.c_[long_1_y,long_1_y_pred,lat_1_y,lat_1_y_pred]).to_csv("unrecontrstructed_prediction_output_1step.csv",index=False)

pd.DataFrame(np.c_[long_multi_y,long_multi_y_pred,lat_multi_y,lat_multi_y_pred]).to_csv("unrecontrstructed_prediction_output_multi_step.csv",index=False)
